Extract resolving currency
